glint-horizon/requirements.txt
# Horizon Core Requirements
Django>=1.11.18
django_compressor>=1.3
django_openstack_auth>=1.1.4
eventlet>=0.13.0
iso8601>=0.1.9
kombu>=2.4.8
lesscpy>=0.9j
# Horizon Utility Requirements
# for SECURE_KEY generation
lockfile>=0.8
netaddr>=0.7.6
pbr>=0.6,!=0.7,<1.0
python-ceilometerclient>=1.0.6
python-cinderclient>=1.0.6
python-glanceclient>=0.9.0
python-heatclient>=0.2.3
python-keystoneclient>=0.7.0
python-neutronclient>=2.3.4,<3
python-novaclient>=2.17.0
python-saharaclient>=0.6.0
python-swiftclient>=1.6
python-troveclient>=1.0.3
pytz>=2010h
six>=1.5.2
